Putting together tests using the emulator

+++ src/storage/storage.spec.ts
@@ -12,7 +12,7 @@ describe('AngularFireStorage', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
+        AngularFireModule.initializeTestApp(COMMON_CONFIG),
         AngularFireStorageModule
       ]
     });
@@ -36,94 +36,97 @@ describe('AngularFireStorage', () => {
 
   it('should have an initialized Firebase app', () => {
     expect(afStorage.storage.app).toBeDefined();
-    // @ts-ignore
-    expect(afStorage.storage.app).toEqual(app);
   });
 
-  describe('upload task', () => {
+  // TODO tests for node?
+  if (typeof Blob !== 'undefined') {
 
-    it('should upload and delete a file', (done) => {
-      const data = { angular: "fire" };
-      const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
-      const ref = afStorage.ref('af.json');
-      const task = ref.put(blob);
-      const sub = task.snapshotChanges()
-        .subscribe(
-          snap => { expect(snap).toBeDefined() },
-          e => { done.fail(); },
-          () => {
-            ref.delete().subscribe(done, done.fail);
-          });
-    });
+    describe('upload task', () => {
 
-    it('should upload a file and observe the download url', (done) => {
-      const data = { angular: "fire" };
-      const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
-      const ref = afStorage.ref('af.json');
-      const task = ref.put(blob).then(() => {;
-        const url$ = ref.getDownloadURL();
-        url$.subscribe(
-          url => { expect(url).toBeDefined(); },
-          e => { done.fail(); },
-          () => { ref.delete().subscribe(done, done.fail); }
-        );
+      it('should upload and delete a file', (done) => {
+        const data = { angular: "fire" };
+        const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
+        const ref = afStorage.ref('af.json');
+        const task = ref.put(blob);
+        const sub = task.snapshotChanges()
+          .subscribe(
+            snap => { expect(snap).toBeDefined() },
+            e => { done.fail(); },
+            () => {
+              ref.delete().subscribe(done, done.fail);
+            });
       });
+
+      it('should upload a file and observe the download url', (done) => {
+        const data = { angular: "fire" };
+        const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
+        const ref = afStorage.ref('af.json');
+        const task = ref.put(blob).then(() => {;
+          const url$ = ref.getDownloadURL();
+          url$.subscribe(
+            url => { expect(url).toBeDefined(); },
+            e => { done.fail(); },
+            () => { ref.delete().subscribe(done, done.fail); }
+          );
+        });
+      });
+
+      it('should resolve the task as a promise', (done) => {
+        const data = { angular: "promise" };
+        const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
+        const ref = afStorage.ref('af.json');
+        const task: AngularFireUploadTask = ref.put(blob);
+        task.then(snap => { 
+          expect(snap).toBeDefined(); 
+          done();
+        }).catch(done.fail);
+      });
+
     });
 
-    it('should resolve the task as a promise', (done) => {
-      const data = { angular: "promise" };
-      const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
-      const ref = afStorage.ref('af.json');
-      const task: AngularFireUploadTask = ref.put(blob);
-      task.then(snap => { 
-        expect(snap).toBeDefined(); 
-        done();
-      }).catch(done.fail);
+    describe('reference', () => {
+
+      it('it should upload, download, and delete', (done) => {
+        const data = { angular: "fire" };
+        const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
+        const ref = afStorage.ref('af.json');
+        const task = ref.put(blob);
+        // Wait for the upload
+        const sub = forkJoin(task.snapshotChanges())
+          .pipe(
+            // get the url download
+            mergeMap(() => ref.getDownloadURL()),
+            // assert the URL
+            tap(url => expect(url).toBeDefined()),
+            // Delete the file
+            mergeMap(url => ref.delete())
+          )
+          // finish the test
+          .subscribe(done, done.fail);
+      });
+
+      it('should upload, get metadata, and delete', (done) => {
+        const data = { angular: "fire" };
+        const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
+        const ref = afStorage.ref('af.json');
+        const task = ref.put(blob, { customMetadata: { blah: 'blah' } });
+        // Wait for the upload
+        const sub = forkJoin(task.snapshotChanges())
+          .pipe(
+            // get the metadata download
+            mergeMap(() => ref.getMetadata()),
+            // assert the URL
+            tap(meta => expect(meta.customMetadata).toEqual({ blah: 'blah' })),
+            // Delete the file
+            mergeMap(meta => ref.delete())
+          )
+          // finish the test
+          .subscribe(done, done.fail);
+      });
+
     });
 
-  });
-
-  describe('reference', () => {
-
-    it('it should upload, download, and delete', (done) => {
-      const data = { angular: "fire" };
-      const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
-      const ref = afStorage.ref('af.json');
-      const task = ref.put(blob);
-      // Wait for the upload
-      const sub = forkJoin(task.snapshotChanges())
-        .pipe(
-          // get the url download
-          mergeMap(() => ref.getDownloadURL()),
-          // assert the URL
-          tap(url => expect(url).toBeDefined()),
-          // Delete the file
-          mergeMap(url => ref.delete())
-        )
-        // finish the test
-        .subscribe(done, done.fail);
-    });
-
-    it('should upload, get metadata, and delete', (done) => {
-      const data = { angular: "fire" };
-      const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
-      const ref = afStorage.ref('af.json');
-      const task = ref.put(blob, { customMetadata: { blah: 'blah' } });
-      // Wait for the upload
-      const sub = forkJoin(task.snapshotChanges())
-        .pipe(
-          // get the metadata download
-          mergeMap(() => ref.getMetadata()),
-          // assert the URL
-          tap(meta => expect(meta.customMetadata).toEqual({ blah: 'blah' })),
-          // Delete the file
-          mergeMap(meta => ref.delete())
-        )
-        // finish the test
-        .subscribe(done, done.fail);
-    });
-
-  });
+  }
 
 });
 
@@ -137,7 +140,7 @@ describe('AngularFireStorage w/options', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
+        AngularFireModule.initializeTestApp(COMMON_CONFIG),
         AngularFireStorageModule
       ],
       providers: [
@@ -168,8 +171,6 @@ describe('AngularFireStorage w/options', () => {
 
     it('should have an initialized Firebase app', () => {
       expect(afStorage.storage.app).toBeDefined();
-      // @ts-ignore
-      expect(afStorage.storage.app).toEqual(app);
     });
 
     it('should be hooked up the right app', () => {
@@ -180,24 +181,29 @@ describe('AngularFireStorage w/options', () => {
       expect(afStorage.storage.ref().toString()).toEqual( `gs://${FIREBASE_STORAGE_BUCKET}/`);
     });
 
-    it('it should upload, download, and delete', (done) => {
-      const data = { angular: "fire" };
-      const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
-      const ref = afStorage.ref('af.json');
-      const task = ref.put(blob);
-      // Wait for the upload
-      const sub = forkJoin(task.snapshotChanges())
-        .pipe(
-          // get the url download
-          mergeMap(() => ref.getDownloadURL()),
-          // assert the URL
-          tap(url => expect(url).toMatch(new RegExp(`https:\\/\\/firebasestorage\\.googleapis\\.com\\/v0\\/b\\/${FIREBASE_STORAGE_BUCKET}\\/o\\/af\\.json`))),
-          // Delete the file
-          mergeMap(url => ref.delete())
-        )
-        // finish the test
-        .subscribe(done, done.fail);
-    });
+    // TODO tests for Node?
+    if (typeof Blob !== 'undefined') {
+
+      it('it should upload, download, and delete', (done) => {
+        const data = { angular: "fire" };
+        const blob = new Blob([JSON.stringify(data)], { type : 'application/json' });
+        const ref = afStorage.ref('af.json');
+        const task = ref.put(blob);
+        // Wait for the upload
+        const sub = forkJoin(task.snapshotChanges())
+          .pipe(
+            // get the url download
+            mergeMap(() => ref.getDownloadURL()),
+            // assert the URL
+            tap(url => expect(url).toMatch(new RegExp(`https:\\/\\/firebasestorage\\.googleapis\\.com\\/v0\\/b\\/${FIREBASE_STORAGE_BUCKET}\\/o\\/af\\.json`))),
+            // Delete the file
+            mergeMap(url => ref.delete())
+          )
+          // finish the test
+          .subscribe(done, done.fail);
+      });
+
+    }
 
   });
